Add Guid.ToString formats
