+++ .editorconfig
@@ -1,0 +1,9 @@
+# To learn more about .editorconfig see https://aka.ms/editorconfigdocs
+
+# All files
+[*]
+indent_style = space
+
+# TS files
+[*.ts]
+indent_size = 2
